Better argument/option parsing. Removed SSH/Telnet functionality.
Not uses optparse to parse arguments/options, which is easier to
read for both devs and users.

Removed SSH/Telnet workers as these depend on libSVM being run
as an external process, which is no longer the case (since we're
using libSVM's python bindings).
Multiple local workers also caused segfaults and errors, so they
were removed.